[refactor] 텍스트 업로드 버튼 위치 헤더로 변경 #173
